OFS-13 Add elasticsearch queries

+++ app/src/main/java/org/openfact/documents/reader/pe/creditnote/PECreditNoteReader.java
@@ -73,7 +73,7 @@ public class PECreditNoteReader implements DocumentReader {
         documentEntity.setCustomerAssignedId(creditNoteType.getAccountingCustomerParty().getCustomerAssignedAccountID().getValue());
         documentEntity.setCustomerName(creditNoteType.getAccountingCustomerParty().getParty().getPartyLegalEntity().get(0).getRegistrationName().getValue());
         documentEntity.setCurrency(creditNoteType.getLegalMonetaryTotal().getPayableAmount().getCurrencyID().value());
-        documentEntity.setAmount(creditNoteType.getLegalMonetaryTotal().getPayableAmount().getValue());
+        documentEntity.setAmount(creditNoteType.getLegalMonetaryTotal().getPayableAmount().getValue().floatValue());
         documentEntity.setIssueDate(creditNoteType.getIssueDate().getValue().toGregorianCalendar().getTime());
 
         Map<String, String> tags = new HashMap<>();
